Cleanup gateways, move process_payment
