+++ modules/ppcp-wc-gateway/src/Gateway/PayPalGateway.php
@@ -9,18 +9,21 @@ declare(strict_types=1);
 
 namespace WooCommerce\PayPalCommerce\WcGateway\Gateway;
 
+use Exception;
 use Psr\Log\LoggerInterface;
-use WooCommerce\PayPalCommerce\ApiClient\Endpoint\OrderEndpoint;
-use WooCommerce\PayPalCommerce\ApiClient\Endpoint\PaymentsEndpoint;
-use WooCommerce\PayPalCommerce\ApiClient\Factory\ShippingPreferenceFactory;
+use WC_Order;
+use WooCommerce\PayPalCommerce\ApiClient\Entity\PaymentToken;
+use WooCommerce\PayPalCommerce\ApiClient\Exception\PayPalApiException;
+use WooCommerce\PayPalCommerce\ApiClient\Exception\RuntimeException;
 use WooCommerce\PayPalCommerce\Onboarding\Environment;
 use WooCommerce\PayPalCommerce\Onboarding\State;
 use WooCommerce\PayPalCommerce\Session\SessionHandler;
+use WooCommerce\PayPalCommerce\Subscription\FreeTrialHandlerTrait;
 use WooCommerce\PayPalCommerce\Subscription\Helper\SubscriptionHelper;
 use WooCommerce\PayPalCommerce\Vaulting\PaymentTokenRepository;
+use WooCommerce\PayPalCommerce\WcGateway\Exception\GatewayGenericException;
 use WooCommerce\PayPalCommerce\WcGateway\FundingSource\FundingSourceRenderer;
 use WooCommerce\PayPalCommerce\WcGateway\Gateway\PayUponInvoice\PayUponInvoiceGateway;
-use WooCommerce\PayPalCommerce\WcGateway\Processor\AuthorizedPaymentsProcessor;
 use WooCommerce\PayPalCommerce\WcGateway\Processor\OrderProcessor;
 use WooCommerce\PayPalCommerce\WcGateway\Processor\RefundProcessor;
 use WooCommerce\PayPalCommerce\WcGateway\Settings\SettingsRenderer;
@@ -32,7 +35,7 @@ use WooCommerce\PayPalCommerce\Webhooks\Status\WebhooksStatusPage;
  */
 class PayPalGateway extends \WC_Payment_Gateway {
 
-	use ProcessPaymentTrait;
+	use ProcessPaymentTrait, FreeTrialHandlerTrait;
 
 	const ID                            = 'ppcp-gateway';
 	const INTENT_META_KEY               = '_ppcp_paypal_intent';
@@ -62,13 +65,6 @@ class PayPalGateway extends \WC_Payment_Gateway {
 	 * @var OrderProcessor
 	 */
 	protected $order_processor;
-
-	/**
-	 * The processor for authorized payments.
-	 *
-	 * @var AuthorizedPaymentsProcessor
-	 */
-	protected $authorized_payments_processor;
 
 	/**
 	 * The settings.
@@ -120,27 +116,6 @@ class PayPalGateway extends \WC_Payment_Gateway {
 	protected $payment_token_repository;
 
 	/**
-	 * The shipping_preference factory.
-	 *
-	 * @var ShippingPreferenceFactory
-	 */
-	private $shipping_preference_factory;
-
-	/**
-	 * The payments endpoint
-	 *
-	 * @var PaymentsEndpoint
-	 */
-	protected $payments_endpoint;
-
-	/**
-	 * The order endpoint.
-	 *
-	 * @var OrderEndpoint
-	 */
-	protected $order_endpoint;
-
-	/**
 	 * Whether the plugin is in onboarded state.
 	 *
 	 * @var bool
@@ -178,30 +153,25 @@ class PayPalGateway extends \WC_Payment_Gateway {
 	/**
 	 * PayPalGateway constructor.
 	 *
-	 * @param SettingsRenderer            $settings_renderer The Settings Renderer.
-	 * @param FundingSourceRenderer       $funding_source_renderer The funding source renderer.
-	 * @param OrderProcessor              $order_processor The Order Processor.
-	 * @param AuthorizedPaymentsProcessor $authorized_payments_processor The Authorized Payments Processor.
-	 * @param ContainerInterface          $config The settings.
-	 * @param SessionHandler              $session_handler The Session Handler.
-	 * @param RefundProcessor             $refund_processor The Refund Processor.
-	 * @param State                       $state The state.
-	 * @param TransactionUrlProvider      $transaction_url_provider Service providing transaction view URL based on order.
-	 * @param SubscriptionHelper          $subscription_helper The subscription helper.
-	 * @param string                      $page_id ID of the current PPCP gateway settings page, or empty if it is not such page.
-	 * @param Environment                 $environment The environment.
-	 * @param PaymentTokenRepository      $payment_token_repository The payment token repository.
-	 * @param ShippingPreferenceFactory   $shipping_preference_factory The shipping_preference factory.
-	 * @param LoggerInterface             $logger  The logger.
-	 * @param PaymentsEndpoint            $payments_endpoint The payments endpoint.
-	 * @param OrderEndpoint               $order_endpoint The order endpoint.
-	 * @param string                      $api_shop_country The api shop country.
+	 * @param SettingsRenderer       $settings_renderer The Settings Renderer.
+	 * @param FundingSourceRenderer  $funding_source_renderer The funding source renderer.
+	 * @param OrderProcessor         $order_processor The Order Processor.
+	 * @param ContainerInterface     $config The settings.
+	 * @param SessionHandler         $session_handler The Session Handler.
+	 * @param RefundProcessor        $refund_processor The Refund Processor.
+	 * @param State                  $state The state.
+	 * @param TransactionUrlProvider $transaction_url_provider Service providing transaction view URL based on order.
+	 * @param SubscriptionHelper     $subscription_helper The subscription helper.
+	 * @param string                 $page_id ID of the current PPCP gateway settings page, or empty if it is not such page.
+	 * @param Environment            $environment The environment.
+	 * @param PaymentTokenRepository $payment_token_repository The payment token repository.
+	 * @param LoggerInterface        $logger  The logger.
+	 * @param string                 $api_shop_country The api shop country.
 	 */
 	public function __construct(
 		SettingsRenderer $settings_renderer,
 		FundingSourceRenderer $funding_source_renderer,
 		OrderProcessor $order_processor,
-		AuthorizedPaymentsProcessor $authorized_payments_processor,
 		ContainerInterface $config,
 		SessionHandler $session_handler,
 		RefundProcessor $refund_processor,
@@ -211,37 +181,25 @@ class PayPalGateway extends \WC_Payment_Gateway {
 		string $page_id,
 		Environment $environment,
 		PaymentTokenRepository $payment_token_repository,
-		ShippingPreferenceFactory $shipping_preference_factory,
 		LoggerInterface $logger,
-		PaymentsEndpoint $payments_endpoint,
-		OrderEndpoint $order_endpoint,
 		string $api_shop_country
 	) {
-
-		$this->id                            = self::ID;
-		$this->order_processor               = $order_processor;
-		$this->authorized_payments_processor = $authorized_payments_processor;
-		$this->settings_renderer             = $settings_renderer;
-		$this->funding_source_renderer       = $funding_source_renderer;
-		$this->config                        = $config;
-		$this->session_handler               = $session_handler;
-		$this->refund_processor              = $refund_processor;
-		$this->transaction_url_provider      = $transaction_url_provider;
-		$this->page_id                       = $page_id;
-		$this->environment                   = $environment;
-		$this->onboarded                     = $state->current_state() === State::STATE_ONBOARDED;
-		$this->id                            = self::ID;
-		$this->order_processor               = $order_processor;
-		$this->authorized_payments           = $authorized_payments_processor;
-		$this->shipping_preference_factory   = $shipping_preference_factory;
-		$this->settings_renderer             = $settings_renderer;
-		$this->config                        = $config;
-		$this->session_handler               = $session_handler;
-		$this->refund_processor              = $refund_processor;
-		$this->transaction_url_provider      = $transaction_url_provider;
-		$this->page_id                       = $page_id;
-		$this->environment                   = $environment;
-		$this->logger                        = $logger;
+		$this->id                       = self::ID;
+		$this->settings_renderer        = $settings_renderer;
+		$this->funding_source_renderer  = $funding_source_renderer;
+		$this->order_processor          = $order_processor;
+		$this->config                   = $config;
+		$this->session_handler          = $session_handler;
+		$this->refund_processor         = $refund_processor;
+		$this->state                    = $state;
+		$this->transaction_url_provider = $transaction_url_provider;
+		$this->subscription_helper      = $subscription_helper;
+		$this->page_id                  = $page_id;
+		$this->environment              = $environment;
+		$this->onboarded                = $state->current_state() === State::STATE_ONBOARDED;
+		$this->payment_token_repository = $payment_token_repository;
+		$this->logger                   = $logger;
+		$this->api_shop_country         = $api_shop_country;
 
 		if ( $this->onboarded ) {
 			$this->supports = array( 'refunds' );
@@ -291,13 +249,6 @@ class PayPalGateway extends \WC_Payment_Gateway {
 				'process_admin_options',
 			)
 		);
-		$this->subscription_helper      = $subscription_helper;
-		$this->payment_token_repository = $payment_token_repository;
-		$this->logger                   = $logger;
-		$this->payments_endpoint        = $payments_endpoint;
-		$this->order_endpoint           = $order_endpoint;
-		$this->state                    = $state;
-		$this->api_shop_country         = $api_shop_country;
 	}
 
 	/**
@@ -306,7 +257,6 @@ class PayPalGateway extends \WC_Payment_Gateway {
 	 * @return bool
 	 */
 	public function needs_setup(): bool {
-
 		return ! $this->onboarded;
 	}
 
@@ -340,9 +290,8 @@ class PayPalGateway extends \WC_Payment_Gateway {
 	 * @return string
 	 */
 	public function generate_ppcp_html(): string {
-
 		ob_start();
-		$this->settings_renderer->render( false );
+		$this->settings_renderer->render();
 		$content = ob_get_contents();
 		ob_end_clean();
 		return $content;
@@ -449,6 +398,126 @@ class PayPalGateway extends \WC_Payment_Gateway {
 			&& self::ID === $this->page_id;
 	}
 	// phpcs:enable WordPress.Security.NonceVerification.Recommended
+
+	/**
+	 * Process payment for a WooCommerce order.
+	 *
+	 * @param int $order_id The WooCommerce order id.
+	 *
+	 * @return array
+	 */
+	public function process_payment( $order_id ) {
+		$wc_order = wc_get_order( $order_id );
+		if ( ! is_a( $wc_order, WC_Order::class ) ) {
+			return $this->handle_payment_failure(
+				null,
+				new GatewayGenericException( new Exception( 'WC order was not found.' ) )
+			);
+		}
+
+		$funding_source = filter_input( INPUT_POST, 'ppcp-funding-source', FILTER_SANITIZE_STRING );
+
+		if ( 'card' !== $funding_source && $this->is_free_trial_order( $wc_order ) ) {
+			$user_id = (int) $wc_order->get_customer_id();
+			$tokens  = $this->payment_token_repository->all_for_user_id( $user_id );
+			if ( ! array_filter(
+				$tokens,
+				function ( PaymentToken $token ): bool {
+					return isset( $token->source()->paypal );
+				}
+			) ) {
+				return $this->handle_payment_failure( $wc_order, new Exception( 'No saved PayPal account.' ) );
+			}
+
+			$wc_order->payment_complete();
+
+			return $this->handle_payment_success( $wc_order );
+		}
+
+		/**
+		 * If customer has chosen change Subscription payment.
+		 */
+		if ( $this->subscription_helper->has_subscription( $order_id ) && $this->subscription_helper->is_subscription_change_payment() ) {
+			$saved_paypal_payment = filter_input( INPUT_POST, 'saved_paypal_payment', FILTER_SANITIZE_STRING );
+			if ( $saved_paypal_payment ) {
+				update_post_meta( $order_id, 'payment_token_id', $saved_paypal_payment );
+
+				return $this->handle_payment_success( $wc_order );
+			}
+		}
+
+		/**
+		 * If the WC_Order is paid through the approved webhook.
+		 */
+		//phpcs:disable WordPress.Security.NonceVerification.Recommended
+		if ( isset( $_REQUEST['ppcp-resume-order'] ) && $wc_order->has_status( 'processing' ) ) {
+			return $this->handle_payment_success( $wc_order );
+		}
+		//phpcs:enable WordPress.Security.NonceVerification.Recommended
+
+		try {
+			if ( ! $this->order_processor->process( $wc_order ) ) {
+				return $this->handle_payment_failure(
+					$wc_order,
+					new Exception(
+						$this->order_processor->last_error()
+					)
+				);
+			}
+
+			if ( $this->subscription_helper->has_subscription( $order_id ) ) {
+				as_schedule_single_action(
+					time() + ( 1 * MINUTE_IN_SECONDS ),
+					'woocommerce_paypal_payments_check_saved_payment',
+					array(
+						'order_id'    => $order_id,
+						'customer_id' => $wc_order->get_customer_id(),
+						'intent'      => $this->config->has( 'intent' ) ? $this->config->get( 'intent' ) : '',
+					)
+				);
+			}
+
+			return $this->handle_payment_success( $wc_order );
+		} catch ( PayPalApiException $error ) {
+			if ( $error->has_detail( 'INSTRUMENT_DECLINED' ) ) {
+				$wc_order->update_status(
+					'failed',
+					__( 'Instrument declined. ', 'woocommerce-paypal-payments' ) . $error->details()[0]->description ?? ''
+				);
+
+				$this->session_handler->increment_insufficient_funding_tries();
+				if ( $this->session_handler->insufficient_funding_tries() >= 3 ) {
+					return $this->handle_payment_failure(
+						null,
+						new Exception(
+							__( 'Please use a different payment method.', 'woocommerce-paypal-payments' ),
+							$error->getCode(),
+							$error
+						)
+					);
+				}
+
+				$host = $this->config->has( 'sandbox_on' ) && $this->config->get( 'sandbox_on' ) ?
+					'https://www.sandbox.paypal.com/' : 'https://www.paypal.com/';
+				$url  = $host . 'checkoutnow?token=' . $this->session_handler->order()->id();
+				return array(
+					'result'   => 'success',
+					'redirect' => $url,
+				);
+			}
+
+			return $this->handle_payment_failure(
+				$wc_order,
+				new Exception(
+					Messages::generic_payment_error_message() . ' ' . $error->getMessage(),
+					$error->getCode(),
+					$error
+				)
+			);
+		} catch ( RuntimeException $error ) {
+			return $this->handle_payment_failure( $wc_order, $error );
+		}
+	}
 
 	/**
 	 * Process refund.
